+++ src/components/LinksPanel/components/LinkElement/LinkElement.tsx
@@ -1,6 +1,6 @@
 import { cx } from '@emotion/css';
 import { Button, Dropdown, LinkButton, MenuItem, Tooltip, useStyles2 } from '@grafana/ui';
-import React from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 
 import { TEST_IDS } from '@/constants';
 import { ButtonSize } from '@/types';
@@ -37,16 +37,52 @@ interface Props {
    * @type {ButtonSize}
    */
   buttonSize?: ButtonSize;
+
+  /**
+   * Dynamic font size
+   *
+   * @type {boolean}
+   */
+  dynamicFontSize?: boolean;
 }
 
 /**
  * Links Element
  */
-export const LinkElement: React.FC<Props> = ({ link, buttonSize, gridMode = false }) => {
+export const LinkElement: React.FC<Props> = ({ link, buttonSize, gridMode = false, dynamicFontSize = false }) => {
   /**
    * Styles
    */
-  const styles = useStyles2(getStyles);
+  const styles = useStyles2(getStyles, { dynamicFontSize });
+
+  /**
+   * States
+   */
+  const [linkWidth, setLinkWidth] = useState(0);
+  const [linkEl, setLinkEl] = useState<HTMLElement | null>(null);
+
+  /**
+   * Get link ref
+   */
+  const btnRef = useCallback((node: HTMLElement | null) => {
+    setLinkEl(node);
+  }, []);
+
+  useEffect(() => {
+    if (!dynamicFontSize || !linkEl) {
+      return;
+    }
+
+    /**
+     * Track link element width changes
+     */
+    const resize = new ResizeObserver(([entry]) => {
+      setLinkWidth(Math.floor(entry.contentRect.width));
+    });
+    resize.observe(linkEl);
+
+    return () => resize.disconnect();
+  }, [dynamicFontSize, linkEl]);
 
   /**
    * Mapping of content alignment positions to their corresponding styles.
@@ -129,13 +165,18 @@ export const LinkElement: React.FC<Props> = ({ link, buttonSize, gridMode = fals
     }
 
     return (
-      <Dropdown
-        key={link.name}
-        overlay={<div className={styles.menu}>{menuLinks}</div>}
-        {...testIds.dropdown.apply(link.name)}
+      <div
+        ref={btnRef}
+        style={dynamicFontSize ? ({ '--btn-width': `${linkWidth}px` } as React.CSSProperties) : undefined}
       >
-        {dropdownButton()}
-      </Dropdown>
+        <Dropdown
+          key={link.name}
+          overlay={<div className={styles.menu}>{menuLinks}</div>}
+          {...testIds.dropdown.apply(link.name)}
+        >
+          {dropdownButton()}
+        </Dropdown>
+      </div>
     );
   }
 
@@ -147,27 +188,32 @@ export const LinkElement: React.FC<Props> = ({ link, buttonSize, gridMode = fals
 
     if (currentLink.url) {
       return (
-        <LinkButton
-          key={currentLink.url}
-          className={cx(
-            currentLink.isCurrentLink ? styles.currentDashboard : styles.link,
-            gridMode && styles.linkGridMode,
-            currentLink.alignContentPosition && alignClassMap[currentLink.alignContentPosition]
-          )}
-          icon={!currentLink.showCustomIcons ? currentLink.icon : undefined}
-          href={currentLink.url}
-          title={currentLink.name}
-          target={currentLink.target}
-          variant="secondary"
-          fill="outline"
-          size={buttonSize}
-          {...testIds.buttonSingleLink.apply(link.name)}
+        <div
+          ref={btnRef}
+          style={dynamicFontSize ? ({ '--btn-width': `${linkWidth}px` } as React.CSSProperties) : undefined}
         >
-          {currentLink.showCustomIcons && currentLink.customIconUrl && !!currentLink.customIconUrl.length && (
-            <img src={currentLink.customIconUrl} alt="" className={styles.customIcon} />
-          )}
-          {currentLink.name}
-        </LinkButton>
+          <LinkButton
+            key={currentLink.url}
+            className={cx(
+              currentLink.isCurrentLink ? styles.currentDashboard : styles.link,
+              gridMode && styles.linkGridMode,
+              currentLink.alignContentPosition && alignClassMap[currentLink.alignContentPosition]
+            )}
+            icon={!currentLink.showCustomIcons ? currentLink.icon : undefined}
+            href={currentLink.url}
+            title={currentLink.name}
+            target={currentLink.target}
+            variant="secondary"
+            fill="outline"
+            size={buttonSize}
+            {...testIds.buttonSingleLink.apply(link.name)}
+          >
+            {currentLink.showCustomIcons && currentLink.customIconUrl && !!currentLink.customIconUrl.length && (
+              <img src={currentLink.customIconUrl} alt="" className={styles.customIcon} />
+            )}
+            {currentLink.name}
+          </LinkButton>
+        </div>
       );
     }
   }
@@ -176,21 +222,26 @@ export const LinkElement: React.FC<Props> = ({ link, buttonSize, gridMode = fals
    * Return
    */
   return (
-    <Button
-      variant="secondary"
-      className={cx(
-        styles.link,
-        gridMode && styles.linkGridMode,
-        link.alignContentPosition && alignClassMap[link.alignContentPosition]
-      )}
-      key={link.name}
-      fill="outline"
-      size={buttonSize}
-      title={link.name}
-      tooltip="Empty URL"
-      {...testIds.buttonEmptyLink.apply(link.name)}
+    <div
+      ref={btnRef}
+      style={dynamicFontSize ? ({ '--btn-width': `${linkWidth}px` } as React.CSSProperties) : undefined}
     >
-      {link.name}
-    </Button>
+      <Button
+        variant="secondary"
+        className={cx(
+          styles.link,
+          gridMode && styles.linkGridMode,
+          link.alignContentPosition && alignClassMap[link.alignContentPosition]
+        )}
+        key={link.name}
+        fill="outline"
+        size={buttonSize}
+        title={link.name}
+        tooltip="Empty URL"
+        {...testIds.buttonEmptyLink.apply(link.name)}
+      >
+        {link.name}
+      </Button>
+    </div>
   );
 };