+++ src/components/editors/GroupsEditor/components/GroupEditor/components/LinkEditor/LinkEditor.tsx
@@ -52,7 +52,7 @@ interface Props extends EditorProps<LinkConfig> {
   data: DataFrame[];
 
   /**
-   * is Grid Mode
+   * Is grid mode
    *
    * @type {boolean}
    */